+++ tests/testthat/test-unnest.R
@@ -101,6 +101,17 @@ test_that("sep combines column names", {
     expect_named(c("x_x", "y_x"))
 })
 
+test_that("can unnest empty data frame", {
+  df <- data_frame(x = integer(), y = list())
+  out <- unnest(df, y)
+  expect_equal(out, data_frame(x = integer()))
+})
+
+test_that("empty ... returns df if no list-cols", {
+  df <- data_frame(x = integer(), y = integer())
+  expect_equal(unnest(df), df)
+})
+
 # Drop --------------------------------------------------------------------
 
 test_that("unnest drops list cols if expanding", {